Make requirements loose for use in projects
When using libraries in larger projects (such as WebPay) that want to pin
requirements, it's hard to work with libs that require exact versions.
This change pins versions to minimums only which will at least give
the developer a hint if something goes wrong (they can revert to the minimums).
Without this change, curling won't run in WebPay because of conflicting
requirements.